ChatGPT Antwort 1
Write a website that offers players information about my minecraft modpack. I want different sections that can be toggled on or off, a table for the modlist the can be sorted in all categories, Some paragraphs and text in the beginning, a nice image placeholder at the top for the page for an artwork. Make the style like in a classic github-readme file (dark mode). I want a section on the left of the screen, where users can see the structure of the document an click around in it. Give me the HTML, CSS and JS files for this website.